# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "API Legal",
    "description": "Plataforma para usuarios individuales y firmas pequeñas del ambito legal, construida sobre una API disenada desde el dia 1 con el rigor de un producto publico.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/auth/registro": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Registro",
        "operationId": "registrar_organizacion",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistroRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistroResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login",
        "operationId": "iniciar_sesion",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/yo": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Yo",
        "operationId": "obtener_actor_actual",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/YoResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LoginRequest": {
        "properties": {
          "organizacion": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Organizacion",
            "description": "Slug de la organización"
          },
          "email": {
            "type": "string",
            "maxLength": 255,
            "minLength": 3,
            "title": "Email"
          },
          "contrasena": {
            "type": "string",
            "maxLength": 72,
            "minLength": 1,
            "title": "Contrasena"
          }
        },
        "type": "object",
        "required": [
          "organizacion",
          "email",
          "contrasena"
        ],
        "title": "LoginRequest"
      },
...
      "RegistroRequest": {
        "properties": {
          "nombre_organizacion": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Nombre Organizacion"
          },
          "nombre": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Nombre"
          },
          "email": {
            "type": "string",
            "maxLength": 255,
            "minLength": 3,
            "title": "Email"
          },
          "contrasena": {
            "type": "string",
            "maxLength": 72,
            "minLength": 8,
            "title": "Contrasena"
          }
        },
        "type": "object",
        "required": [
          "nombre_organizacion",
          "nombre",
          "email",
          "contrasena"
        ],
        "title": "RegistroRequest"
      },
      "RegistroResponse": {
        "properties": {
          "organizacion_id": {
            "type": "string",
            "format": "uuid",
            "title": "Organizacion Id"
          },
          "organizacion_slug": {
            "type": "string",
            "title": "Organizacion Slug"
          },
          "usuario_id": {
            "type": "string",
            "format": "uuid",
            "title": "Usuario Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          }
        },
        "type": "object",
        "required": [
          "organizacion_id",
          "organizacion_slug",
          "usuario_id",
          "email"
        ],
        "title": "RegistroResponse"
      },
...
      "TokenResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "expira_en_segundos": {
            "type": "integer",
            "title": "Expira En Segundos"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "expira_en_segundos"
        ],
        "title": "TokenResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "YoResponse": {
        "properties": {
          "usuario_id": {
            "type": "string",
            "format": "uuid",
            "title": "Usuario Id"
          },
          "organizacion_id": {
            "type": "string",
            "format": "uuid",
            "title": "Organizacion Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          }
        },
        "type": "object",
        "required": [
          "usuario_id",
          "organizacion_id",
          "email"
        ],
        "title": "YoResponse"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}